Checking video in folder and deleting file if download error

+++ downloader.py
@@ -161,9 +161,21 @@ def main():
                 print("{}".format(el['msg']))
 
             filename = os.path.join(folder_name, 'Video_' + str(week) + '.mp4')
-            print('Downloading file ', filename)
-            urllib.request.urlretrieve(el['url'], filename)
-            print('Done')
+            if not os.path.isfile(filename):
+                try:
+                    print('Downloading file ', filename)
+                    urllib.request.urlretrieve(el['url'], filename)
+                    print('Done')
+                except urllib.error.ContentTooShortError:
+                    os.remove(filename)
+                    print('Error while downloading. File {} deleted:'.format(filename))
+                except KeyboardInterrupt:
+                    if os.path.isfile(filename):
+                        os.remove(filename)
+                    print('\nAborted')
+                    exit(1)
+            else:
+                print('File {} already exist'.format(filename))
         print("All steps downloaded")
